+++ packages/brand-components/src/Avatar/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import s from './styles.module.scss';
 import cx from 'classnames';
-import { VisualTheme } from '../../../gamut/src/theming/VisualTheme';
+import { VisualTheme } from '@codecademy/gamut';
 
 type AvatarImageProps =
   | {
@@ -10,7 +10,7 @@ type AvatarImageProps =
     }
   | { alt?: never; 'aria-labelledby': string };
 
-export type AvatarBaseProps = {
+type AvatarBaseProps = {
   src: string;
   size?: 'regular' | 'large';
   theme?: VisualTheme;