+++ packages/brand-components/CHANGELOG.md
@@ -3,6 +3,79 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.1](https://github.com/Codecademy/client-modules/compare/@codecademy/brand-components@0.3.0...@codecademy/brand-components@0.3.1) (2020-02-25)
+
+
+### Bug Fixes
+
+* testimonial component background pattern ([1a08cd2](https://github.com/Codecademy/client-modules/commit/1a08cd2ddc84c59cc1088dfdd62e37afff452a36))
+
+
+
+
+
+# [0.3.0](https://github.com/Codecademy/client-modules/compare/@codecademy/brand-components@0.2.4...@codecademy/brand-components@0.3.0) (2020-02-24)
+
+
+### Features
+
+* **testimonial:** create testimonial brand component in size small ([9a765c8](https://github.com/Codecademy/client-modules/commit/9a765c847548b952c0f8c2343a1b646f7a2d009b))
+
+
+
+
+
+## [0.2.4](https://github.com/Codecademy/client-modules/compare/@codecademy/brand-components@0.2.3...@codecademy/brand-components@0.2.4) (2020-02-24)
+
+**Note:** Version bump only for package @codecademy/brand-components
+
+
+
+
+
+## [0.2.3](https://github.com/Codecademy/client-modules/compare/@codecademy/brand-components@0.2.2...@codecademy/brand-components@0.2.3) (2020-02-24)
+
+**Note:** Version bump only for package @codecademy/brand-components
+
+
+
+
+
+## [0.2.2](https://github.com/Codecademy/client-modules/compare/@codecademy/brand-components@0.2.1...@codecademy/brand-components@0.2.2) (2020-02-24)
+
+**Note:** Version bump only for package @codecademy/brand-components
+
+
+
+
+
+## [0.2.1](https://github.com/Codecademy/client-modules/compare/@codecademy/brand-components@0.2.0...@codecademy/brand-components@0.2.1) (2020-02-21)
+
+**Note:** Version bump only for package @codecademy/brand-components
+
+
+
+
+
+# [0.2.0](https://github.com/Codecademy/client-modules/compare/@codecademy/brand-components@0.1.1...@codecademy/brand-components@0.2.0) (2020-02-21)
+
+
+### Features
+
+* create Byline  brand component  ([e330622](https://github.com/Codecademy/client-modules/commit/e330622f8e9780a2ed49b411339a310f3d1b049e))
+
+
+
+
+
+## [0.1.1](https://github.com/Codecademy/client-modules/compare/@codecademy/brand-components@0.1.0...@codecademy/brand-components@0.1.1) (2020-02-20)
+
+**Note:** Version bump only for package @codecademy/brand-components
+
+
+
+
+
 # 0.1.0 (2020-02-20)