+++ packages/gamut-templates/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.35](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-templates@2.2.34...@codecademy/gamut-templates@2.2.35) (2020-02-24)
+
+**Note:** Version bump only for package @codecademy/gamut-templates
+
+
+
+
+
+## [2.2.34](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-templates@2.2.33...@codecademy/gamut-templates@2.2.34) (2020-02-24)
+
+**Note:** Version bump only for package @codecademy/gamut-templates
+
+
+
+
+
+## [2.2.33](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-templates@2.2.32...@codecademy/gamut-templates@2.2.33) (2020-02-24)
+
+**Note:** Version bump only for package @codecademy/gamut-templates
+
+
+
+
+
+## [2.2.32](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-templates@2.2.31...@codecademy/gamut-templates@2.2.32) (2020-02-20)
+
+**Note:** Version bump only for package @codecademy/gamut-templates
+
+
+
+
+
 ## [2.2.31](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-templates@2.2.30...@codecademy/gamut-templates@2.2.31) (2020-02-17)
 
 **Note:** Version bump only for package @codecademy/gamut-templates